Add parts types.
The input SDL statement will be splitted in different `queryParts`. We
support three classes:

 * PassthroughPart
 * InputPart
 * OutputPart

Input and Output parts provide constructors and accessors for
convenience. The constructors check for the correctness of the
parameters and returns an error if the expressions are malformed.

The definition of the `QueryPart` interface has been moved to the
`parts.go` file and has been made private.