Add Prepare stage for inputs (#28)
This PR contains the Prepare stage for input parts. In the prepare stage the user passes instantiations of every Go struct that is mentioned in an input or output expression. These objects are used only for information about their type, the actual values inside the structs are irrelevant.

The prepareInput function checks that the structs and tags mentioned in IO expressions in the query correspond to the types of the objects passed as arguments to Prepare.