refactor: include expression in prepare errors (#99)
In prepare.go, include the I/O expression in which the error was found where possible. The errors are changed to reflect that the full expression is now included in the error message. The phrase "cannot prepare expression" is changed to "cannot prepare statement" to reflect that expression is used to refer only to the I/O expression.