Change parsed number from string to pointer
We need to represent that the low and high indexes are optional, we were
using strings for that with the empty string being the missing value,
problem was that we had to parse the number eventually. By using
pointers we get the same type safety of having to check the actual value
to see if it is not empty, while also not having to parse. The only
caveat is that we could cause panics if we are not careful
de-referencing.