refactor: better error on missing output columns (#89)
Add an error triggered when columns from output expressions do not appear in query results. This could happen when a output expressions is within a WITH clause of a query.

Also refactor TestIterGetErrors to be more concise and reliable. Before, if no error was thrown by iter.Get the test could still pass.