Change names and add documentation for new pipeline (#122)
Change type/variable names and add documentation for the new SQLair pipeline. A `doc.go` file is added to the `expr` package to detail how the pipeline should look. The names of the types in the pipeline are changed to clarify their purpose.

The full planned changes are described in the `doc.go` file added here. The upcoming changes In short are: the SQL method will be on `PrimedQuery` rather than `TypedExpr` and the `TypedExpr` will be changed to retain the shape of the AST from the `ParsedExpr`.