Colocate methods with type definitions (#123)
Move major types to separate files along with their methods and helper functions. Also move (internally) public methods on the types to the top of the files.

Before, the files were organised by the stage in the pipeline. This change moves the methods to be co-located with the type definitions rather than ordering them according to the pipeline.

Methods are only moved here, no code is changed. The documentation in `internal/expr/doc.go` is updated to reflect the new file names.