Test for prepared statement leaks (#119)
Created wrapper driver to monitor statement creation and deletion while calling SQLite behind the scenes.

Before this change the tests did not flag when statements were not closed properly, for example if you commented [the following line](https://github.com/canonical/sqlair/blob/main/sqlair.go#L279) no tests would have failed. After this change, the tests are more robust and they can now verify that the cache and finalizers are working correctly. This is especially important now that we want to overhaul the cache implementation and we want to verify it is still correct.