refactor: remove cache (#128)
There are several issues including:
* Overhead of finalizers.
* Limitation in the cache design when the max number of connections is
  1, because of the sql standard library works.
* It is not clear whether preparing statements in SQLite shows an
  improvement in performance.
which require time investment and through testing to qualify performance
gains. Right now, we are going to focus on delivering the missing
features for adoption of SQLair. That is why we have decided to remove
the cache and revisit the conversation in the future when we can commit
with the necessary investment.