bugfix: Fix duplicate output error (#129)
Fix error where a single map member can be used as an output multiple times in a query.

The `outputUsed` map mapped `typeinfo.Output` to booleans recording if the output had been used. For maps the `mapKey` object that implemented the `typeinfo.Output` has a pointer receiver. It was possible to have two identical `mapKey` structs with different pointers.

Here, the error is fixed by indexing `outputUsed` with a string representation of the `typeinfo.Output` using its `String` method.