Add support for parsing unicode chars in column names and tags (#140)
Add compatibility with unicode

The parser is changed to check runes rather than bytes, it also allows for numbers and quoted names to be parsed as column names. The validation for tag names is also changed to include the above so that these columns can be parsed directly into structs.

You will now be able to write SQLair queries such as:

SELECT 世界 AS &S.* FROM t
SELECT 1 AS &M.doesExist FROM t
SELECT "string or column name!" AS $M.quoted FROM t

Also tidied up use of go check in some of the tests, and added tests for lower case structs.