Add a cache for driver prepared statements.
When a sqlair statement is run directly against the database it is first
prepared on the driver and then exectued. The driver prepared statement
is saved in a sqlair cache and reused. When a statement or DB goes out
of memory then its associated driver preparaed statements are evicted
from the cache.