Refactor bindtypes into typedexprbuilder and check unused prepared args
This refactor adds a typedExprBuilder that handles getting the
information about the arguments and aggregating the typed expressions.
It also keeps track of which arguments have been used to allow for
useful error messages.

This is nearly all a refactor excpet for a single new error message
which is triggered when extra arguements are passed to prepare