feat: add a cache for driver prepared statements (#149)
* Add a cache for driver prepared statements.

When the same SQLair query is executed multiple times against the database it is recompiled by the database every time. This is inefficient.

It is possible to instead prepare a SQL query against the database driver and cache it in SQLair to allow for reuse of the query.

With this change, when a SQLair statement is executed directly against a database it is first prepared and then executed, the prepared statement is put in the cache for reuse. If the statement is executed in a transaction, the SQLair statement is looked up in the cache but if it is not found it is not prepared on the driver. This is due to the issues in https://github.com/canonical/sqlair/pull/117.

* Allow a single statement to correspond to multiple SQL strings in cache

* SQUASH Review corrections