docs: move syntax docs to RTD reference
The syntax docs were originaly part of the top level doc.go. The syntax
should be moved to the read the docs because it can be better expressed
there and we can have a single source of reference.

The doc.go file can now be deleted as it is obsolete. On pkg.go.dev the
github readme is displayed there which links to the RTD docs.