fix: remove close DB on sqlair.DB going out of memory
As part of the cache work, a finalizer was put on the sqlair.DB that
would close the underlying sql.DB when the sqlair.DB went out of memory.

This was a misfeature, if thread seperatly held the sql.DB or the same
thread tried to use it later on it would be inexplicably closed.

Remove the db.Close on the sqlair.DB finalizer.