fix: memory leak
When statements are eviceted from the cache, they are not always closed.
This change fixes that.

Some sqlair.Statements can have multiple sql queries generated from them
depending on the arguements (e.g. bulk inserts, or queries with slice
arguements). SQLair maintains a sqlair.Statement to driverStmt cache for
DB queries. When one of these queries is executed, and it is looked up
in the cache, is a driver statement is found, but has different SQL, it
will be removed from the cache and replaced with the newly generated SQL
from the version of the query with different arguments.

Before this change, a finalizer was set on the *sql.Stmt to close it
when it dropped from memory. The *sql.Stmt couldn't be closed straight
away becuase there is a possibility someone else is using it
concurrently, therefor the finalizer was set to clear it up after it had
been finished with.

This didn't work as intended. In fact, the sql.DB holds a reference to
this *sql.Stmt so the finalizer will only ever be run when this is
closed. The test designed to check that this statement was closed did
not pick it up becuase before the check, the DB was closed. In a recent
change, the DB stopped being closed at this point, and this caused the
test to start failing.

After this change, instead of putting a finalizer on the database/sql
controller *sql.Stmt, we put a finalizer on the SQLair controller
driverStmt. This means we can audit usages far more effectivly and are
not dependent on third part code.

If the finalizer is run before all concurrent users of the *sql.Stmt
have finished with it, sql.Rows may return an error saying query closed.
This happens when the underlying statement is closed. To ensure that the
finalizer is only run when all concuirrent users have finished with it,
the Iterator will now hold a reference to this driverStmt. That way,
it will only be once the Iterator goes out of memory that the driverStmt
is closed.