fix: panic on no columns
SQLair would panic when it expected to find at least one column in the
results and found none. It would create a slice using the length of the
found columns and then try and access it with the expected index of the
column.

An error is now thrown in this case instead or panic, this error is also
tested.

Part of the reason this panic occured is becuase the ScanArgs function
is undertested. It had no unit tests, only the black box tests in
package_test.go. Unit tests for errors have been added in expr_test,
though it could still do with more explict happy path tests.