build: update to go version 1.24
Update to the latest supported go version, 1.24. Go 1.18 is unmaintained
and thus a security risk.

According to pkg.go.dev, SQLair is imported by the open source projects
github.com/juju/juju and github.com/canonical/notary. Both of these are
on go version 1.24 so this should not cause them any issues when
updating to the latest version of SQLair.