refactor: do not track transaction done status
There is no need to monitor the done status of Sqlair transactions. This
is already done in the standard lib transaction implementation at a
level where there is negligible perfomance benefit.

Most importantly, a context cancellation automatically rolls back a
transaction, which marks it done in the standard implementation, but of
which Sqlair would be unaware.