refactor: do not track transaction done status (#191)
There is no need to monitor the done status of Sqlair transactions. This is already done in the standard lib transaction implementation at a level where there is negligible performance benefit to adding the tracking at a layer above.

More importantly, a context cancellation automatically rolls back a transaction, which marks it done in the standard implementation, but of which Sqlair would be unaware. In such a case the work to track the status is made redundant.

Here we simply remove the tracking. The modified test asserts that the behaviour is effectively unchanged.